Fixed bug in CodeIgniter's captcha helper
Fixed bug in CodeIgniter's captcha helper. It now only removes image-files in
the captcha folder. This explained the mysterious deletion of
`images/captcha/folder-info.txt`.

Note that strict comparison is used, !(strpos($filename, ".jpg") === false)).
It's not really necessary, but it would cause problesm is ".jpg" was a valid
filename, as it would return 0 as the position, and would evaluate to false.

  system/helpers/captcha_helper.php